requirements.txt
# =====================================================
# MCP Analytics Server - Phase 2 Dependencies
# Python 3.11+
# =====================================================

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.34.0
gunicorn==21.2.0

# MCP Protocol
fastmcp>=2.12.5,<3

# Database - PostgreSQL
psycopg2-binary==2.9.11
SQLAlchemy==2.0.25

# ORM & Validation
pydantic>=2.11.7
pydantic-settings>=2.1.0

# Redis (Cache + Pub/Sub)
redis==5.0.1
hiredis==2.3.2

# Background Tasks (Celery)
celery==5.3.6
kombu==5.3.5

# LLM Integration
openai>=1.50.0

# Security (Connection string encryption)
cryptography==42.0.0

# SQL Parsing & Validation
sqlparse==0.5.3

# HTTP Client
httpx==0.28.1

# Utilities
python-dotenv>=1.1.0
python-dateutil==2.8.2

